Add new XDG_STATE_HOME variable
The basedir spec somewhat recently added XDG_STATE_HOME to account for things like history that are not cache, and not temporary, but alsö not config and alsö not data. The history file itself may need to last quite a long time, and arguably should still be backed up (not cache), but it is constantly in flux and would make zero sense in version control (not config).